step3: seperate Direction from Point.kt
