+++ src/main/kotlin/minesweeper/Direction.kt
@@ -1,0 +1,12 @@
+package minesweeper
+
+enum class Direction(val row: Int, val col: Int) {
+    NORTH(-1, 0),
+    NORTHEAST(-1, 1),
+    EAST(0, 1),
+    SOUTHEAST(1, 1),
+    SOUTH(1, 0),
+    SOUTHWEST(1, -1),
+    WEST(0, -1),
+    NORTHWEST(-1, -1)
+}
